[ renderd ] Bug fix 32 bit / 64 bit issue
When calculating the the intermediate hashkey I incorrectly assumed long was 64 bit. This meant that the hashkey could turn negative
which in turn would cause a segfault on array access.

Thanks to Richard Ive for spotting and debugging this error.



git-svn-id: http://svn.openstreetmap.org/applications/utils/mod_tile@17439 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ daemon.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <unistd.h>
 #include <sys/types.h>
 #include <sys/socket.h>
@@ -121,12 +122,12 @@ void clear_requests(int fd)
 
 
 static int calcHashKey(struct item *item) {
-    int xmlnameHash = 0;
-    long key;
+    uint64_t xmlnameHash = 0;
+    uint64_t key;
     for (int i = 0; item->req.xmlname[i] != 0; i++) {
         xmlnameHash += item->req.xmlname[i];
     }
-    key = ((long)(xmlnameHash & 0x1FF) << 52) + ((long)(item->req.z) << 48) + ((long)(item->mx & 0xFFFFFF) << 24) + (item->my & 0xFFFFFF);
+    key = ((uint64_t)(xmlnameHash & 0x1FF) << 52) + ((uint64_t)(item->req.z) << 48) + ((uint64_t)(item->mx & 0xFFFFFF) << 24) + (item->my & 0xFFFFFF);
     return key % hashidxSize;
 }
